backend/requirements.txt
fastapi
uvicorn
pdfplumber
python-multipart

# AI providers
google-generativeai
openai
anthropic